+++ test/operators.jl
@@ -100,6 +100,7 @@ end
     @test  isunordered(NaN)
     @test  isunordered(NaN32)
     @test  isunordered(missing)
+    @test  isunordered(nothing)
     @test !isunordered(1)
     @test !isunordered([NaN, 1])
     @test !isunordered([1.0, missing])
@@ -281,3 +282,15 @@ end
 end
 
 @test [Base.afoldl(+, 1:i...) for i = 1:40] == [i * (i + 1) ÷ 2 for i = 1:40]
+
+# pr #40031
+@test isless(0, nothing)
+@test !isless(nothing, 0)
+@test !isless(nothing, nothing)
+@test isless(missing, nothing)
+@test !isless(nothing, missing)
+@test_throws MethodError 0 < nothing
+@test_throws MethodError nothing < 0
+@test_throws MethodError nothing < nothing
+@test_throws MethodError missing < nothing
+@test_throws MethodError nothing < missing